api handlers: do not set TlsVerify
Commit 52a4642edd8a2c2f62d10c2180d785b4f04f18c5 forced the default to
`true` in the code along with the docs.  The problem with hard-setting
it is that it breaks how the optional bool is meant to be used by
containers/image.  If it is not set, c/image will default to true unless
otherwise configured in `/etc/containers/registries.conf`.  If set,
c/image will use the specified value as it assumes the _user_ knows what
they are doing.

In other words: without this fix, the API ignores the `insecure`
settings in `registries.conf`.

[NO NEW TESTS NEEDED]

Signed-off-by: Valentin Rothberg <vrothberg@redhat.com>