Fixes #15206 Correctly handle ssh to machine when VM found for WSL
On Windows the VM is not correctly identified when the connection is not
default. The load VM by name passes, but wants to try a remote
connection which results in a no service configured.

This change will not check the validVM flag but instead if the load
returned a valid entry.

[NO NEW TESTS NEEDED]

Signed-off-by: Gerard Braad <me@gbraad.nl>