Man pages: refactor common options: --device-X-Y
In all cases, went with the version from podman-run because
it conforms to our documentation guidelines. I then added
the FAQ-26 link to each one.

This does not review easily with hack/markdown-preprocess-review
because there are too many files involved. Sorry.

This will cause conflicts in #15276 - again, I'm sorry. On the bright
side, that PR still needs work anyway, and rebasing on top of
this should make it easier to review.

Signed-off-by: Ed Santiago <santiago@redhat.com>