Cirrus: Don't run win_installer in multiarch cron
The win_installer task fails on the `multiarch` cirrus-cron build.
This is because it depends on the `Windows Cross` (alt_build) task
which is bypassed in this context. This will cause the `repo.tbz`
download to constantly throw 404s.  Fix this by skipping the
win_installer task for the `multiarch` (container images) build.

Signed-off-by: Chris Evich <cevich@redhat.com>